Edit comments, file names and code based on PR comments
