+++ src/thunderbots/software/CMakeLists.txt
@@ -652,12 +652,12 @@ if (CATKIN_ENABLE_TESTING)
 
     target_link_libraries(timestamp_test ${catkin_LIBRARIES})
 
-    catkin_add_gtest(evaluation_player_orientation_within_angle_threshold_of_target_test
-            test/evaluation/player.cpp
-            ai/hl/stp/evaluation/player.cpp
+    catkin_add_gtest(evaluation_robot_orientation_within_angle_threshold_of_target_test
+            test/evaluation/robot.cpp
+            ai/hl/stp/evaluation/robot.cpp
             )
 
-    target_link_libraries(evaluation_player_orientation_within_angle_threshold_of_target_test ${catkin_LIBRARIES})
+    target_link_libraries(evaluation_robot_orientation_within_angle_threshold_of_target_test ${catkin_LIBRARIES})
